Churn_Modelling_Spark_Machine_Learning_Classification
Business Problem:

This data set contains details of a bank's customers and the target variable is a binary variable reflecting the fact whether the customer left the bank (closed his account) or he continues to be a customer.

Goal:

It is our duty to predict  the customer left the bank (closed his account) or he continues to be a customer by using spark machine learning classification.